Sort array of objects with identical keys by their values
When a property value is an array of objects with identical keys,
optionally sort the objects by the value of the keys. This can be
particularly interesting for autogenerated files that are updated by an
external application. i.e. in amplify the configuration files are
serverside updated. Diffing of these files is really hard when the order
changes. A lot of thes files contain arrays of objects with identical
keys. Sorting them by value will always keep them in the same order.